Expressions with literal value

+++ core/src/main/kotlin/verik/core/it/ItExpressionExtractor.kt
@@ -16,11 +16,11 @@
 
 package verik.core.it
 
+import verik.core.base.LineException
 import verik.core.it.symbol.ItSymbolTable
 import verik.core.lang.Lang
 import verik.core.lang.LangFunctionExtractorRequest
 import verik.core.lang.LangSymbol.TYPE_INT
-import verik.core.base.LineException
 import verik.core.sv.SvExpression
 import verik.core.sv.SvExpressionFunction
 import verik.core.sv.SvExpressionLiteral
@@ -64,7 +64,7 @@ object ItExpressionExtractor {
         val typeReified = literal.typeReified
                 ?: throw LineException("literal expression has not been reified", literal)
         val string = when (typeReified.type) {
-            TYPE_INT -> literal.value.toString()
+            TYPE_INT -> literal.value.toInt().toString()
             else -> throw LineException("literal type not supported", literal)
         }
         return SvExpressionLiteral(
